==> 1. pdf_textract.ipynb <==
### 1.3 pyPDF4 reading pdf files.

+ get a list of files from file directory filedir

from os import listdir
from os.path import isfile,join
from tika import parser

import pickle

+ use Tika package parser to read and extract pdf contents and save them in contentDICT.
+ need JAVA8 installed

def dirPDF2pkl(filedir,filename):
    minsFiles = [f for f in listdir(filedir) if isfile(join(filedir,f))]
    contentDICT = {f:parser.from_file(join(filedir,f))['content'] for f in minsFiles}
    with open(filename, 'wb') as handle:
        pickle.dump(contentDICT, handle, protocol=pickle.HIGHEST_PROTOCOL)
        
    return contentDICT

+ save the files contents dictionary with their file name

fileDIR = './data'
filename = "council_minutes_dict.pkl"

content = dirPDF2pkl(fileDIR,filename)

+ validate pickle file read.

with open(filename,'rb') as file:
    check = pickle.load(file)
type(check)
