==> 99. dump.ipynb <==
def longestSubstringFinder(string1, string2):
    answer = ""
    len1, len2 = len(string1), len(string2)
    for i in range(len1):
        match = ""
        for j in range(len2):
            if (i + j < len1 and string1[i + j] == string2[j]):
                match += string2[j]
            else:
                if (len(match) > len(answer)): answer = match
                match = ""
    return answer

news = '2.4'
word_ls = re.split(' ',news)
word_ls = [w for w in word_ls if w!=' ']
word_ls = [w for w in word_ls if w!='']
word_ls = [w for w in word_ls if w!='\n']\

word_ls
i = 2
try:
    main = float(word_ls[0])
    if math.floor(main)==i:
        j = (main-i)*10
        print('I am here')
except Exception as err:
    print('I should not be here')
    print(err)
    
j

import regex as re
import math

i = 2

testS = '2.1 Approval of City Council Minutes.'

word_ls = re.split(' ',testS)
word_ls = [w for w in word_ls if w!=' ']
word_ls = [w for w in word_ls if w!='']
word_ls = [w for w in word_ls if w!='\n']

try:
    main = float(word_ls[0])
    if math.floor(main)==i:
        j = (main-i)*10
        print('I am here')
except Exception as err:
    print('I should not be here')
    print(err)

class article:
    
    def __init__(self,name,body='',patterns = {}):
        self.name = name
        self.body = body
        self.patterns = self.validate_patterns(patterns)
    
    def set_name(self,new_name):
        self.name = new_name
        
    def set_body(self,new_body):
        self.body = new_body
        
    def set_body(self,new_patterns):
        self.patterns = new_patterns
        
    def validate_patterns(self,patterns):
        if sum([type(p) ==type('') for p in patterns])==0:
            return patterns
        else:
            return None
        
    def 

### extract content as per pattern.

maintitlechk = {t:[] for t in pattern_dict['main title']}

for (key, byte_article) in raw.items():
    article = byte_article.decode('ISO8859-1')
    
    for title in pattern_dict['main title']:
        
        if title in article:
            maintitlechk[title].append(key)
        else:
            pass
            
maintitlechk

maintitlechk = {t:0 for t in pattern_dict['main title']}

for (key, byte_article) in raw.items():
    article = byte_article.decode('ISO8859-1')
    
    for title in pattern_dict['main title']:
        
        if title in article:
            maintitlechk[title]+=1
        else:
            pass
            
maintitlechk

def lsExtract(rawLS):
    extracted = rawLS[0]
    for ls in rawLS:
        for item in extracted:
            if item in ls:
                pass
            else:
                extracted.remove(item)
    
    return extracted

maintitle = lsExtract(maintitle)
maintitle

# following content is not used at this point.

maintitle = re.findall(pattern_maintitle,min_01)
subtitle = re.findall(pattern_subtitle,min_01)
item = re.findall(pattern_item,min_01)
dollar = re.findall(pattern_dollar,min_01)

print(maintitle)
#print(subtitle)
#print(item)
#print(dollar)

common words

# Creat sents and words for listed sentances and words.

from nltk.tokenize import sent_tokenize, word_tokenize

sents = sent_tokenize(min_01)
words = word_tokenize(min_01)

words_lower = [w.lower() for w in words]
words_unique = set(words_lower)

# create stop words list and remove stop words from word list

from nltk.corpus import stopwords

english_stop = stopwords.words('english')

words_alpha = [t for t in words_lower if t.isalpha()] #extracta only alphabetic words
words_ns = [w for w in words_alpha if w not in english_stop]

from nltk.stem import WordNetLemmatizer

word_lemmatizer = WordNetLemmatizer()

words_lemm = [word_lemmatizer.lemmatize(w) for w in words_ns]

from collections import Counter

bow_s = Counter(words_lemm)
print(bow_s.most_common(20))

Extract naming entity

sents_word = [word_tokenize(s) for s in sents]

from gensim.corpora.dictionary import Dictionary

dictionary = Dictionary([words])

gen_corpus = dictionary.doc2bow(words)

bow_gen_corpus = sorted(gen_corpus,key=lambda w:w[1],reverse=True)

for word_id, word_count in bow_gen_corpus[:5]:
    print(dictionary.get(word_id), word_count)

from collections import defaultdict
import itertools

total_word_count = defaultdict(int)

for word_id, word_count in gen_corpus:
    total_word_count[word_id] += word_count

from gensim.models.tfidfmodel import TfidfModel

tfidf = TfidfModel([gen_corpus])
tfidf_weights = tfidf[gen_corpus]

from nltk import pos_tag,ne_chunk_sents

sents_pos = [pos_tag(s) for s in sents_word]

sents_chunked = ne_chunk_sents(sents_pos,binary=True)
chunk_ls = []

for sent in sents_chunked:
    for chunk in sent:
        if hasattr(chunk,'label'):
            chunk_ls.append(chunk)

chk_ls=[]
for chunk in chunk_ls:
    for c in chunk:
        chk_ls.append(c)

new_ls =[]

for chk in chk_ls:
    (a,b) = chk
    a = a.lower()
    new_ls.append((a,b))
    
new_ls[4] == new_ls[0]

def countLS(LS):
    sum_ls = []
    for i in LS:
        temp_ls = [a for [a,b] in sum_ls]
        if i in temp_ls:
            sum_ls[temp_ls.index(i)][1] += 1
        else:
            sum_ls.append([i,1])
    
    return sum_ls

sum_ls = countLS(new_ls)

sum_ls

testString = ['$5.00',
              '$50',
              '$5,000.00',
              '$5000',
              '$5,000', 
              '$5,000,000.00',
              '$5,000,000,000',
              '$5,000,000']

import regex as re

pattern = re.compile('([$]\d+([,.]\d+)*)')

for s in testString:
    sub = re.findall(pattern,s)
    if sub is not None:
        print(sub,'pass')
    else:
        print(sub,'failed')

sectionpattern = re.compile(r'\s((Sections?)\s\d+[.]\d+[.]\d+)')

s = '(a) Adopt a resolution amending the 2019-2020 Schedule of Fees and Charges \n(Resolution No. 72737, as amended) to reduce plan review, building permit, and impact \nfees to $0 for property owners who participate in the ADU Amnesty Program and who \nqualify for a Financial Hardship Exemption under the criteria set forth in San José \nMunicipal Code Sections 4.76.400 through 4.76.470.\n(b) Adopt a resolution amending the Schedule of Parkland Fees (Resolution No. 73587, \nas amended) to reduce fees to $0 for property owners who participate in the ADU \nAmnesty Program and who qualify for a Financial Hardship Exemption under the \ncriteria set forth in San José Municipal Code Sections 4.76.400 through 4.76.470.\nCEQA: Not a Project, File No. PP17-005, Adjustment to Fees, Rates and Fares without \nchanges to or expansion of services. (Planning, Building and Code Enforcement/Parks, \nRecreation, and Neighborhood Services)\n'

sub = re.findall(sectionpattern,s)
sub

## 1. Amazon Textract

import json
import pandas as pd

# read file
with open('raw/city of san jose minutes_10.json', 'r') as myfile:
    data=myfile.read()

# parse file
obj = json.loads(data)

def creatRaw(result):

    raw_ls=[]
    raw_txt=''
    for t in result['Blocks']:
        try:
#             if t['Text'][:4]=='Page':
#                 pass
#             else:
            raw_ls.append( t['Text'])
            raw_txt += t['Text']+' '
        except:
            pass

    return raw_txt,raw_ls

creatRaw(obj)

## 2. Textract

import pickle 

import textract
text = textract.process('data/Minutes_10.pdf', method='pdfminer')

str_text = text.decode('UTF-8')

clean_text = str_text.replace(("\n" or "\n\n" or "\x0c")," ")

# Read all the text into one string
print(str_text)

file_head = 'admin/city_council_minutes/Minutes_'
file_suffix = '.pdf'
target_head = 'min_'

d = dict()

for i in range(1,10):
    d[str(i)] = textract.process((file_head + str(0) + str(i) + file_suffix), method='pdfminer')

for i in range(10,46):
    d[str(i)] = textract.process((file_head + str(i) + file_suffix), method='pdfminer')


def save_obj(obj, name ):
    with open('data/'+ name + '.pkl', 'wb') as f:
        pickle.dump(obj, f, pickle.HIGHEST_PROTOCOL)

def load_obj(name ):
    with open('data/' + name + '.pkl', 'rb') as f:
        return pickle.load(f)

save_obj(d, "council_minutes_dict")

t = load_obj("council_minutes_dict")
